pyproject.toml
[tool.poetry]
name = "international-delivery"
version = "0.1.0"
description = ""
authors = ["muhycspb <dimka.k@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
asyncpg = {extras = ["fastapi"], version = "^0.29.0"}
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
uvicorn = "^0.30.5"
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
dynaconf = "^3.2.6"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
